requirements.txt
alembic==1.4.1
anyio==2.2.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
catboost==0.25.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
coloredlogs==15.0
cycler==0.10.0
databricks-cli==0.14.3
dataclasses-json==0.5.4
decorator==5.0.7
defusedxml==0.7.1
deprecation==2.1.0
docker==5.0.0
entrypoints==0.3
fasttext==0.9.2
Flask==1.1.2
funcy==1.15
gitdb==4.0.7
GitPython==3.1.15
graphviz==0.16
greenlet==1.0.0
gunicorn==20.1.0
humanfriendly==9.1
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json-logging==1.3.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.9.1
jupyter-server==1.6.2
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lightgbm==3.2.1
llvmlite==0.36.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.12.1
marshmallow-enum==1.5.1
matplotlib==3.4.1
missingno==0.4.2
mistune==0.8.4
mlflow==1.15.0
mypy-extensions==0.4.3
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numba==0.53.1
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.1
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pyarrow==4.0.0
pybind11==2.6.2
pycparser==2.20
pydantic==1.8.2
Pygments==2.8.1
pynndescent==0.5.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
querystring-parser==1.2.4
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smmap==4.0.0
sniffio==1.2.0
SQLAlchemy==1.4.10
sqlparse==0.4.1
stringcase==1.2.0
tabulate==0.8.9
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tomlkit==0.7.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
typing-inspect==0.7.1
umap-learn==0.5.1
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
